Fix issue with setting hashes when length exceeds threshold
Having the conditional statement to check if the
length is greater than or equal to the threshold means
that any time the set method is called (regardless if
we are setting a new hash or overwriting an old one),
the rehash method is called and we expand the number of
buckets. This is not the intended effect. Thus, checking
for the length after the setting is done solves this.