Fix issue with setting hashes when length is checked if exceeds threshold
The code before had the conditional statement at the
bottom of the method and the value of the length of the
hashmap stoerd at the top. This means that we were
comparing the length before the hash was set with the
threshold, which is incorrect. Moving the top statement
down and just above the conditional check solves this.
Additionally, making sure the check is > and not >=
solves the problem of expanding the buckets size too early.