Fix issue with displaying length of hashmap when hashes are removed from one element linked lists
Initially, the issue was believed to be in the LinkedList's
removeAt method. This is because whenever the method is run
on linked lists with only one element, the 'next' element
would be assigned to the head. This obviously does not work
because there is only one element in the list, so the new head
reads null. This was fine for just linked list applications,
but using them in the HashMap object causes the program to
throw errors when these null elements still exist and are
trying to be read as elements in the linked list. Setting the
bucket to null and removing the linked list (when it's empty)
solves this.