Fix issue with automatic resizing when removing elements
The original code for automatic resizing in the remove
method made it so that anytime an element was removed
and the length of the hashmap was less than half of the
capacity times the load factor, it would reduce the
size of the buckets. However, doing this meant that
you'd eventually reduce the size of the buckets down
multiple times and when it isn't needed (the length of
the hashmap is well below the threshold) just by removing
a single element. Changing the condition such that the length
has to equal the threshold solves this.